requirements.txt
flask
google-generativeai
fpdf
python-dotenv
